requirements.txt
# python-dotenv
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.28.0
